refactor(services): merge_service 를 패키지로 승격해 한도 안에 넣는다 (B5.5 Task 3)
`services/merge_service.py` 1,328 줄 → 패키지 3 모듈, 최대 707 줄.
`from services.merge_service import get_merge_service` 가 글자 그대로 유효하게
남아 프로덕션 호출부(`api/git/merge.py` · `api/git/_shared.py`)를 고치지 않는다.

    errors.py      7줄  MergeServiceError
    service.py   707줄  try/except ImportError + MergeService + get_merge_service
    requests.py  633줄  _merge_requests + MergeRequestService

**패치 타깃은 한 이름 6 건 — `GIT_AVAILABLE`.**
`try: from git import ... / except ImportError:` 가 세우는 플래그이고 읽는 곳은
`MergeService.__init__` 하나뿐이다(`GitCommandError` 를 읽는 네 곳도 전부
`MergeService` 안). try 문은 원자 단위라 세 이름이 통째로 그 모듈로 간다.
**배럴은 `GIT_AVAILABLE` 을 재노출하지 않는다** — 재노출하면 낡은 패치가
성공하되 무효가 된다. 실측: 재지정 전 6 건이 `@patch` 지점에서 AttributeError
로 실패(RED), 재지정 후 15 통과.

**`try/except ImportError` 는 fail-open 이라 별도 단언이 필요하다.**
순환 import 가 생기면 "의존성 부재" 로 삼켜져 조용히 저하되고 ruff·mypy·본문
대조·전체 스위트가 **전부 통과한 채로** 그렇게 된다. 실측으로 확인했다 —
폴백을 강제 발동시키니 새 단언 하나만 실패하고 나머지 12 건은 초록이었다.
`test_git_import_did_not_fall_back` 이 그 층을 덮는다(값으로 판정).

**`_merge_requests` 는 `global` 문이 없지만 그것이 판정 근거가 아니다.**
`split_usage.py` 의 `_codex_plan_cache` 는 첨자 대입만 하는데 테스트가 dict 를
통째로 갈아끼웠다. 전수 grep 으로 재바인딩 0 건 · reader 전부
`MergeRequestService` 안임을 확인하고 그 클래스와 같은 모듈에 뒀다.

**배럴이 `MergeRequestStatus`(models.git 소유)를 재노출하는 것은 tmux 와 다르다.**
테스트 3 곳이 실제로 여기서 가져가고, 패치 타깃이 아니라 재노출이 무효를 만들
여지가 없다. 기준은 "정의했는가" 가 아니라 "재바인딩되는가" 다.

**검증** — split_audit 정의 8 종 보존, 분할 전후 실행 비교(플래그·저장소 공유·
배럴 표면·팩토리) JSON diff 완전 일치, 계약 테스트 13 건(배럴 재노출/의도적
누락/폴백/패치 도달/저장소 공유)에 Red-Green 2 종, ruff · ruff format · mypy(323) ·
pytest 1,711 passed / 0 failed.
