+++ src/backend/services/merge_service/errors.py
@@ -1,0 +1,7 @@
+"""머지 서비스 예외."""
+
+
+class MergeServiceError(Exception):
+    """Merge service specific error."""
+
+    pass
